feat: ✨ Extract libboot
