+++ libs/boot/.cargo/config.toml
@@ -1,0 +1,2 @@
+[target.aarch64-metta-none-eabi]
+rustflags = ["-C", "relocation-model=pie"]
