+++ libs/boot/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "libcpu"
+name = "libboot"
 
-description = "Vesper nanokernel CPU support."
+description = "Kernel boot code"
 
 authors = { workspace = true }
 categories = { workspace = true }
@@ -15,28 +15,14 @@ version = { workspace = true }
 
 publish = false
 
-[package.metadata.layer]
-deps = []
-description = ""
-
 [badges]
 maintenance = { status = "experimental" }
 
 [dependencies]
 aarch64-cpu = { workspace = true }
-bit_field = { workspace = true }
-bitflags = { workspace = true }
-cfg-if = { workspace = true }
-once_cell = { workspace = true }
-snafu = { workspace = true }
+libcpu = { workspace = true }
+libplatform = { workspace = true }
 tock-registers = { workspace = true }
-usize_conversions = { workspace = true }
-ux = { workspace = true }
-
-[lib]
-test = false
-
-# For proper testing in libcpu, we build it as a test_runner binary!
 
 [lints]
 workspace = true
