+++ libs/boot/src/arch/aarch64/mod.rs
@@ -1,0 +1,1 @@
+pub mod boot;
