+++ libs/boot/src/arch/mod.rs
@@ -1,0 +1,1 @@
+pub mod aarch64;
