+++ libs/boot/src/lib.rs
@@ -1,0 +1,4 @@
+#![no_std]
+#![no_main]
+
+pub mod arch;
