--- libs/cpu/src/cpu/boot.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: MIT OR Apache-2.0
-//
-// Copyright (c) 2021-2022 Andre Richter <andre.o.richter@gmail.com>
-
-//! Boot code.
-
-// Not used, arch/../cpu/boot.rs is used directly to generate boot code.
-// #[cfg(target_arch = "aarch64")]
-// #[path = "../arch/aarch64/cpu/boot.rs"]
-// mod arch_boot;
